Add unit tests for redaction
